Add sample code for the loop
Use pipes for communication, and fork off a process so that
we run both initiator and acceptor.

The idea is to have the sample code be compilable as-is, with no
dangling references.  My testing should get this to be runnable, too,
though that will require some setting of the target name which will
not end up in the spec.

The native OS X GSSAPI library also doesn't expose a useful
GSS_C_NT_USER_NAME without going through contortions -- the header
in /usr/include/gssapi/ declares a symbol by that name but the
header from the actual implementation, in
/System/Library/Frameworks/GSS.Framework/Headers/ , has that
as a preprocessor macro.

This commit builds but does not run.