A slightly different approach to freeing output_token
It is always safe to release the output token from a context
establishment call, per 5.26 of RFC 2744.  Do it unconditionally
within the loop, therefore.

However, applications are only *required* to free the storage when
the returned token is "non-NULL" (RFC 2743), aka nonzero-length
(RFC 2744).  Since gss_release_buffer() must set the length to
zero, check the length instead of the value, and get rid of the
need to set a separate flag condition.